mac80211: mesh: factor out common plink close code
Reject and accepted close events always put the host in the
holding state and send a close frame to the peer.  Do this in
a single place to save some duplicated code.  Drop the reason
field from the caller as it is no longer needed.  Also, use the
sta->plid value instead of plid local since it is valid for
all close cases.

Signed-off-by: Bob Copeland <me@bobcopeland.com>